Consolidated memlib generate script
